+++ src/MicroOcpp/Model/Reset/ResetService.h
@@ -1,13 +1,16 @@
 // matth-x/MicroOcpp
-// Copyright Matthias Akstaller 2019 - 2023
+// Copyright Matthias Akstaller 2019 - 2024
 // MIT License
 
-#ifndef RESETSERVICE_H
-#define RESETSERVICE_H
+#ifndef MO_RESETSERVICE_H
+#define MO_RESETSERVICE_H
 
 #include <functional>
+#include <vector>
 
+#include <MicroOcpp/Version.h>
 #include <MicroOcpp/Core/ConfigurationKeyValue.h>
+#include <MicroOcpp/Model/Reset/ResetDefs.h>
 
 namespace MicroOcpp {
 
@@ -52,5 +55,60 @@ std::function<void(bool isHard)> makeDefaultResetFn();
 }
 
 #endif //MO_PLATFORM == MO_PLATFORM_ARDUINO && (defined(ESP32) || defined(ESP8266))
+
+#if MO_ENABLE_V201
+
+namespace MicroOcpp {
+
+class Variable;
+
+namespace Ocpp201 {
+
+class ResetService {
+private:
+    Context& context;
+
+    struct Evse {
+        Context& context;
+        ResetService& resetService;
+        const unsigned int evseId;
+
+        std::function<bool(ResetType type)> notifyReset; //notify firmware about a Reset command. Return true if Reset is okay; false if Reset cannot be executed
+        std::function<bool()> executeReset; //execute Reset of connector. Return true if Reset will be executed; false if there is a failure to Reset
+
+        unsigned int outstandingResetRetries = 0; //0 = do not reset device
+        unsigned long t_resetRetry;
+
+        bool awaitTxStop = false;
+
+        Evse(Context& context, ResetService& resetService, unsigned int evseId);
+
+        void loop();
+    };
+
+    std::vector<Evse> evses;
+    Evse *getEvse(unsigned int connectorId);
+    Evse *getOrCreateEvse(unsigned int connectorId);
+
+    Variable *resetRetriesInt = nullptr;
+
+public:
+    ResetService(Context& context);
+    ~ResetService();
+    
+    void loop();
+
+    void setNotifyReset(std::function<bool(ResetType)> notifyReset, unsigned int evseId = 0);
+    std::function<bool(ResetType)> getNotifyReset(unsigned int evseId = 0);
+
+    void setExecuteReset(std::function<bool()> executeReset, unsigned int evseId = 0);
+    std::function<bool()> getExecuteReset(unsigned int evseId = 0);
+
+    ResetStatus initiateReset(ResetType type, unsigned int evseId = 0);
+};
+
+} //namespace Ocpp201
+} //namespace MicroOcpp
+#endif //MO_ENABLE_V201
 
 #endif