+++ CHANGELOG.md
@@ -23,6 +23,7 @@
     - Transactions (preview only), UCs E01 - E12 ([#247](https://github.com/matth-x/MicroOcpp/pull/247))
     - StatusNotification compatibility ([#247](https://github.com/matth-x/MicroOcpp/pull/247))
     - ChangeAvailability compatibility ([#285](https://github.com/matth-x/MicroOcpp/pull/285))
+    - Reset compatibility, UCs B11 - B12 ([#286](https://github.com/matth-x/MicroOcpp/pull/286))
 
 ### Fixed